:truck: Rename server.h into server_data.h
