+++ server/src/data/load_team.c
@@ -8,7 +8,7 @@
 
 #include <stdlib.h>
 #include <string.h>
-#include "server.h"
+#include "server_data.h"
 
 void deserialize_team(roundtable_server_t *server, json_object_t *team)
 {
